1. Performance Optimization 2. Code Splitting
